Start implementing single precision
Full matrix assembly should be abloe to use single precision to
some degree. This will never be fully single precision as
python/numpy/numba always default to double precision
(like if something is multiplied by 2...)

+++ fe2ms/assembly_nonsingular_full.py
@@ -248,9 +248,9 @@ def assemble_B_integral(
                 cols.append(facet2edge[facet, i_edge_m])
                 vals.append(-integral * jacobian)
 
-    rows_array = _np.zeros(len(vals), dtype=_np.int64)
-    cols_array = _np.zeros(len(vals), dtype=_np.int64)
-    B_array = _np.zeros(len(vals))
+    rows_array = _np.zeros(len(vals), dtype=_np.int32)
+    cols_array = _np.zeros(len(vals), dtype=_np.int32)
+    B_array = _np.zeros(len(vals), dtype=basis.dtype)
     for i in range(len(rows)): # pylint: disable=consider-using-enumerate
         rows_array[i] = rows[i]
         cols_array[i] = cols[i]
